0.8.0 refactor + changed inputs to config file
Refactored into a more pythonic format; it's now "config"-file based
(see the README for details eventually) and part of that also allows
for sprintf-like formatting of the value part of the label as well as
being able to specify the theme colors in the standard RGB hex string
format. The code now assumes LABEL,COL1VAL,COL2VAL as the CSV input in
preparation for an arbitrary # of columns (though I suspect 4 might be
the most useful max). Without hacking the code directly, it still only
supports PDF output and you'll have to remove the background fill code
if you want a transparent background. Those options are coming.